issue: 352235 bug fix - define size of MsgPool as long and not int

+++ c/src/MsgPool.h
@@ -37,7 +37,7 @@ public:
 
 	bool error_creating;
 	char   			*buf;
-	int 			buf_size;
+	long 			buf_size;
 	Msg** 			msg_ptrs;
 
 private:
